local option disable zsh ssh agent

+++ .zshrc
@@ -1,11 +1,13 @@
 [ -f ~/.localenv ] && source ~/.localenv
 
-SSH_AGENT_ENV=$HOME/.config/ssh-agent.env
-source $SSH_AGENT_ENV > /dev/null
+if [ -z "$LOCALENV_NOAGENT" ] ; then
+    SSH_AGENT_ENV=$HOME/.config/ssh-agent.env
+    source $SSH_AGENT_ENV > /dev/null
 
-if [ -z "$SSH_AGENT_PID" ] || ! pgrep -u `whoami` -x ssh-agent > /dev/null ; then
-    pgrep -u `whoami` -x ssh-agent | xargs kill -KILL 2&>1 > /dev/null
-    eval $(ssh-agent | tee $SSH_AGENT_ENV)
+    if [ -z "$SSH_AGENT_PID" ] || ! pgrep -u `whoami` -x ssh-agent > /dev/null ; then
+        pgrep -u `whoami` -x ssh-agent | xargs kill -KILL 2&>1 > /dev/null
+        eval $(ssh-agent | tee $SSH_AGENT_ENV)
+    fi
 fi
 
 alias kty='i3-msg "exec --no-startup-id kitty -1 --instance-group i3 group0 -d $(pwd)"'
